refactor(onboard): extract onboard-inference-probes.ts
Move inference endpoint probe functions (parseJsonObject,
hasResponsesToolCall, shouldRequireResponsesToolCalling,
getProbeAuthMode, getValidationProbeCurlArgs, probeResponsesToolCalling,
probeOpenAiLikeEndpoint, probeAnthropicEndpoint) into a dedicated
module.

All dependencies (http-probe, credentials, platform, validation) are
imported directly — no circular deps. The 4 validate* wrapper functions
remain in onboard.ts for now as they depend on promptValidationRecovery;
they'll consolidate when setupNim moves in a later step.

Update source-scanning tests to read from the new compiled file.
