+++ test/wsl2-probe-timeout.test.ts
@@ -42,8 +42,9 @@ describe("WSL2 inference verification timeouts (issue #987)", () => {
     // The retry logic is embedded in probeOpenAiLikeEndpoint which is not
     // exported. Verify the retry triggers on the correct curl exit codes by
     // scanning the compiled source for the guard condition.
+    // probeOpenAiLikeEndpoint moved to onboard-inference-probes.ts
     const onboardSrc = fs.readFileSync(
-      path.join(import.meta.dirname, "..", "dist", "lib", "onboard.js"),
+      path.join(import.meta.dirname, "..", "dist", "lib", "onboard-inference-probes.js"),
       "utf-8",
     );
